+++ lib/ggity/geom/line.ex
@@ -1,7 +1,7 @@
 defmodule GGity.Geom.Line do
   @moduledoc false
 
-  alias GGity.{Draw, Geom, Scale}
+  alias GGity.{Draw, Geom, Labels, Scale}
 
   @type t() :: %__MODULE__{}
   @type record() :: map()
@@ -227,7 +227,7 @@ defmodule GGity.Geom.Line do
         stroke_width: "0.5",
         transform: "translate(#{interval}, 0)"
       ),
-      Draw.text(format_tick(geom_line.x_scale, value),
+      Draw.text(Labels.format(geom_line.x_scale, value),
         fill: "gray",
         y: "9",
         dy: "0.71em",
@@ -264,25 +264,13 @@ defmodule GGity.Geom.Line do
         stroke_width: "0.5",
         transform: "translate(0, -#{interval})"
       ),
-      Draw.text(Float.to_string(value), fill: "gray", x: "-9", dy: "0.32em", font_size: "8")
+      Draw.text(Labels.format(geom_line.y_scale, value),
+        fill: "gray",
+        x: "-9",
+        dy: "0.32em",
+        font_size: "8"
+      )
     ]
     |> Draw.g(opacity: "1", transform: "translate(0,#{top_shift - coord})")
   end
-
-  defp format_tick(%scale_type{date_labels: ""}, value)
-       when scale_type in [Scale.X.Date, Scale.X.DateTime] do
-    to_string(value)
-  end
-
-  defp format_tick(%scale_type{date_labels: {pattern, options}}, value)
-       when scale_type in [Scale.X.Date, Scale.X.DateTime] do
-    NimbleStrftime.format(value, pattern, options)
-  end
-
-  defp format_tick(%scale_type{date_labels: pattern}, value)
-       when scale_type in [Scale.X.Date, Scale.X.DateTime] and is_binary(pattern) do
-    NimbleStrftime.format(value, pattern)
-  end
-
-  defp format_tick(_scale, value), do: to_string(value)
 end