+++ test/ggity_labels_test.exs
@@ -1,0 +1,5 @@
+defmodule GGityLabelsTest do
+  use ExUnit.Case
+
+  doctest GGity.Labels
+end